// src/app/shared/store/types/packaging.type.ts
export enum PackagingActionTypes {
  GET_PACKAGING_DROPDOWN_OPTIONS_SUCCESS = 'GET_PACKAGING_DROPDOWN_OPTIONS_SUCCESS',
  GET_PACKAGING_DROPDOWN_OPTIONS_FAILURE = 'GET_PACKAGING_DROPDOWN_OPTIONS_FAILURE',

  GET_PACKAGING_TYPES = 'GET_PACKAGING_TYPES',

  GET_PACKAGING_OPTIONS = 'GET_PACKAGING_OPTIONS',

  GET_OUTER_PACKAGINGS = 'GET_OUTER_PACKAGINGS',

  UPDATE_PACKAGING_MODEL = 'UPDATE_PACKAGING_MODEL',

  ADD_NEW_PACKAGE = 'ADD_NEW_PACKAGE',

  RESET_PACKAGING_STORE = 'RESET_PACKAGING_STORE',
  RESET_PACKAGING_STORE_BASED_ON_CONCENTRATION = 'RESET_PACKAGING_STORE_BASED_ON_CONCENTRATION',
  RESET_PACKAGING_CHILDREN = 'RESET_PACKAGING_CHILDREN',

  SET_PRODUCT_TYPE = 'SET_PRODUCT_TYPE',

  GET_FILL_VOLUME_VALID_RANGE = 'GET_FILL_VOLUME_VALID_RANGE',
  GET_FILL_VOLUME_VALID_RANGE_SUCCESS = 'GET_FILL_VOLUME_VALID_RANGE_SUCCESS',
  GET_FILL_VOLUME_VALID_RANGE_FAILURE = 'GET_FILL_VOLUME_VALID_RANGE_FAILURE',

  CLEAR_STATE = 'CLEAR_STATE'
}